python-worker/requirements.txt
backoff==2.2.1
Deprecated==1.2.14
googleapis-common-protos==1.61.0
grpcio==1.59.0
importlib-metadata==6.8.0
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
protobuf==4.24.4
temporalio==1.3.0
types-protobuf==4.24.0.2
typing_extensions==4.8.0
wrapt==1.15.0
zipp==3.17.0
